Fixed a bunch of things, primarily a bug related to graph editing: incident edges were not deleted along with their endpoints; turns out that I forgot to add them to the adjacency lists when parsing the graph.

+++ src/edu/ncsu/csc/Galant/graph/component/GraphElement.java
@@ -259,8 +259,12 @@ public class GraphElement {
         return getDouble(state, WEIGHT);
     }
 
-	public void setWeight(Double weight) {
-        set(WEIGHT, weight);
+    /**
+     * Need to convert the argument to a double because it eventually
+     * converts to Double -- see cleanupAfterParsing()
+     */
+	public void setWeight(double weight) {
+        set(WEIGHT, (double) weight);
     }
 
     public boolean hasWeight() {
@@ -405,4 +409,4 @@ public class GraphElement {
     }
 }
 
-//  [Last modified: 2015 07 27 at 19:27:50 GMT]
+//  [Last modified: 2015 08 07 at 19:11:20 GMT]
